added model factory
